Created outline of edge detection tests
These tests are run on a struct that has paths to the input and expected
result along with the parameters used to get to that result

Testing asserts that the edge detection image is only black or white
pixels, that the dimensions of the input and the expected result match
the actual result, and that the difference in individual pixels between
the actual and expected is less than some delta.

We sill need to actually write the methods that count the different
pixels, as well as the method to make sure the dimentions are the same.