Simply edge splitting code
We don't actually need to do all of that connected component mapping to
split the segments on edges. Now we just split when we encounter the
end of an edge, as marked from going from a non-zero pixel to a zero
pixel.

This also changes the signature for creating segments with edge
mapping, because at the moment we're just going to create the edge map
from the original image. Therefore, we don't need an additional edge
map paramter