Implement filter draw method
This implements the filter's draw method which should allow us to
actually perform pixel sorts with arbitrary sort functions.

As part of this we define a new type for sort functions that operate on
colors instead of raw RGBA values. We also define what it means to sort
by a colorSortFunction using the sorting interface