Add function to generate a point order, with test
This adds a function (with a test) that generates what we call a 'point
order'. This is an array of arrays of points. We should be able to use
this both to generate arrays of arrays of colors (that can then be
sorted), as well as to reconstruct the image when the points are sorted.