Implement angled sorting using bresenham's line algo
This implements angled sorting using Bresenham's line algorithm.
The idea is that we use the sort angle to compute a line that we can
then increment in a particular direction to generate angled,
non-overlapping segments of the image, that we then sort.

This only currently works between the angles of 270 and 315 degrees,
but will be modified to work for all angles.